Updated xen_bus_dmamap_load().
It is done to be in sync with the bus_dma_tag_create() implementation.
xen_bus_dmamap_load() now takes an array of grant references as
argument. This array was allocated by xen_bus_dma_tag_create() and it
contains the grant references corresponding to the frames we will load
with the dma map.

The callback method is also updated. It now uses the references array
and updates the frame number using the bus segment's ds_addr.