Fixed not being able to call client's callback in xen_bus_dmamap_load()
In the earlier implementation, we were not calling the callback
specified by the client because we were passing the refs array to the
xen-specific callback and had no way to pass the function pointer and
the arg.

Now, we use a struct to hold the function pointer, client callback
argument and the xen callback argument.