Updated xen_bus_dma_tag_create() to conform to the busdma interface
Earlier, I implemented xen_bus_dma_tag_create() as a helper on top of
the existing busdma implementation. This was not the exact goal of the
project. Changed the header to conform to the dma tag creation interface

Also, grant entries are not granted here. Rather they will be granted
on load.