The grant references are now allocated in tag create, not map_complete
Earlier, the grant references were being allocated in map_complete. This
meant that if the allocation failed, there was no way to return the
error code, or even signal an error. The error was silent and the client
would never know.

Now the grant references are allocated in tag create. Here, error codes
can be returned. Only mapping happens in map_complete now, which can not
fail.