Allocation of grant references now happen for each map, not each tag
A single tag can be used to create multiple maps. So allocating grant
references for the tag, and not for each map is wrong. It also means we
"over-allocate" grant refs. Grant refs are an scarce resource and should
be reserved for as little time as possible.

To achieve this, a new struct bus_dmamap_xen is used. Addition of the
xen-specific map meant every method using dma maps had to be refactored.