Fixed grant refs not getting allocated from deferred context
When the load gets deferred, the _load() functions do not allocate the
grant references. So these must be allocated from the deferred context.

To achieve this, we need to pass a custom callback to the parent which
first allocates and maps the grant references and only then calls the
client's callback.