Failing to allocate grant references is no longer fatal
If the allocation of grant references fails, we now request a callback
to be notified when the shortage is over.

This change introduced multiple complications:

- If the load is deferred, and then the allocation of grant refs is also
deferred, some ugly hacks are needed to make it work.

- If the allocation of grant refs is deferred, and an unload is called
in the meantime, we will probably cause a segmentation fault and leak
grant refs.

- Probably a lot more scenarios which I haven't accounted for where we
introduce bugs. The code has gotten rather complicated becuase of this.
The client was returned EINPROGRESS upon load when the allocation of
grant references failed. So the callback must be called from here.