Restructured the code paths when there is a need to call the xen_gnttab_free_callback()
Earlier, there were two possible ways to reach the
xen_gnttab_free_callback(): One from the load. The other from the
deferred context when the load is deferred. Both the two was had to be
handled differently.

Now, we restructure in a way that there is no need for different
handling of those two cases. This is achieved by calling the parent's
map_complete() a little early (when in a non-deferred context) and
saving a copy of the segs array.

Now all the free callback has to do is allocate the refs, map them and
call client callback, making it much simpler.