Fixed passing wrong address to memcpy() in xbd_queue_cb()
The array of indirection refs has to be copied to the ring so it can use 
the indirection pages.

Before the update to use the new busdma implementation, 
cm_indirectionrefs was an array. So passing &cm->cm_indirectionrefs was 
OK. But now, it as a pointer instead. So &cm->cm_indirectionrefs would 
result in a wrong memory location being copied, and thus the operations 
using the indirection refs would fail.

This patch fixes the disk errors printing on the screen every once in a 
while.