Fixed a memory leak and double free in the blkfront driver
cm->cm_sg_refs was earlier allocated and managed by the blkfront code. 
Now it is done by the busdma code.

So we should not allocate cm_sg_refs because we set it in xbd_queue_cb() 
to the dma map's refs array. This allocation will leak and never get 
freed.

Also, since cm_sg_refs is handled by the busdma interface, there is no 
need to free it either. If we free it, it frees the map's refs array 
(because it was set in xbd_queue_cb()). So it results in a double free 
when we unload the map.