Added the ability to reserve grant references on map creation
This would be needed by netfront because it can't fail when mapping 
grant refs, so the load should not fail. Pre-allocating grant refs makes 
a failed load very unlikely. Not that it can still fail when there is a 
shortage of bounce pages, but that is very unlikely to happen.

The number of grant references reserved is the maximum number of 
segments possible, so this should be used judicially because grant 
references are a scarce resource.