Changed loop limit from pool_idx to NET_RX_RING_SIZE in disconnect_rxq()
We want to make sure we destroy every map. It might be possible that 
there are some maps out of the pool that won't get destroyed if we run 
the loop till pool_idx.

If any of the maps are still loaded, the assertions in the busdma 
functions will help find about any potential bugs.