Changed bus_dmamap_load_mbuf() to bus_dmamap_load()
load_mbuf() loads the entire mbuf chain. But we are doing that manually, 
so we don't want the entire chain to get loaded, just the current mbuf.

Also, load_mbuf() loads an already-loaded map when there more than one 
mbuf in the chain. Since the busdma implementation does not support this 
yet, it can lead to some side-effects.