Moved the DMA tag from one per device instance to per rx/tx queue
Using the same tag for rx and tx queues needs synchronization so that it 
is not used by both at the same time causing data corruption.

Not only is this synchronization difficult to implement, it would 
probably decrease performance as well. So, create a DMA tag for each rx 
or tx queue. Accesses to these queues are already synchronized, so it 
would not lead to any subtle bugs.

The changed related to tx queues are coming in the following commit.