Imporved failure handling in setup_rxqs() and setup_txqs()
Earlier, they would probably cause a panic or page fault by accessing 
invalid memory locations because everything was not set up properly, and 
disconnect_txq()/disconnect_rxq() did not know this.

Here, we clean up for the failed queue in the function, and then call 
disconnect for the rest of the queues.