Split xen_load_helper() into two functions
It was pretty large and had too many checks for the different number of
states it can be called from.

For example, the NOLOAD option is rather unclear about what it does and
why it is there. So, remove the NOLOAD option. Instead, split the work
of xen_load_helper() into two functions so xen_dmamap_callback() can
call that part directly without having to jump through the "NOLOAD"
hoops.