Implemented the Xen-specific bus_dma(9) grant table handlers
These grant table handlers are a transparent replacement of the generic busdma handlers for Xen drivers. This allows for a more transparent integration with the rest of the OS.

busdma_xen.c: The xen-specific busdma implementation resides in this
file.

files.amd64: Added an entry for busdma_xen.c

files.i386: Added an entry for busdma_xen.c